requirements.txt
# Векторные клиенты
qdrant-client>=1.7.0
pymilvus>=2.3.0
psycopg[binary]>=3.1.0
pgvector>=0.2.5

# Эмбеддинги и данные
sentence-transformers>=2.3.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# Метрики и отчёты
scikit-learn>=1.3.0
tabulate>=0.9.0
matplotlib>=3.7.0
ragas>=0.1.0 

# Утилиты
pyyaml>=6.0
python-dotenv>=1.0.0